+++ src/main/java/com/find/doongji/search/payload/response/RecommendResponse.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class RecommendResponse {
-    private int danjiId;
+    private Long danjiId;
     private float similarity;
 
     @Override
